first pass export of upstream request from sparc
This is a first pass at a pull request generated by a semi-automated
workflow from InterLex and the SPARC term request pipeline. This is a
starting point for figuring out the next steps for being able to
transition to a fully automated pull request pipeline. One of those next
steps will be to reserve an id block for requests coming from
tgbugs/NIFSTD/InterLex. We also need to review what information is
currently missing from what we collect as part of the SPARC pipeline.